RH-32-Make_endpoint_for_password_reset_request
Deleted PasswordRequestController and moved its methods in AuthController, renamed inner classes in PasswordRequestValidator
