+++ src/main/java/sevensmurfs/rehub/controller/AuthController.java
@@ -13,10 +13,13 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import sevensmurfs.rehub.model.entity.RehubUser;
+import sevensmurfs.rehub.model.message.request.PasswordResetRequest;
 import sevensmurfs.rehub.model.message.request.UserRequest;
+import sevensmurfs.rehub.model.message.request.validator.PasswordResetValidator;
 import sevensmurfs.rehub.model.message.request.validator.UserRequestValidator;
 import sevensmurfs.rehub.model.message.response.UserResponse;
 import sevensmurfs.rehub.security.JwtGenerator;
+import sevensmurfs.rehub.service.PasswordResetService;
 import sevensmurfs.rehub.service.UserService;
 import sevensmurfs.rehub.util.SecurityUtil;
 
@@ -31,6 +34,8 @@ public class AuthController {
     private final JwtGenerator jwtGenerator;
 
     private final UserService userService;
+
+    private final PasswordResetService passwordResetService;
 
     /**
      * User login request POST > /api/v1/auth/login
@@ -53,5 +58,37 @@ public class AuthController {
         log.info(" < < < POST /api/v1/auth/login (User login request successful)");
 
         return ResponseEntity.ok().body(UserResponse.mapAuthenticatedUserEntity(user, jwtToken));
+    }
+
+    /**
+     * Request password reset link POST > /api/v1/auth/reset
+     */
+    @PostMapping("/reset")
+    public ResponseEntity<Object> getPasswordResetLink(@Validated(PasswordResetValidator.getPasswordResetLink.class)
+                                                       @RequestBody PasswordResetRequest passwordResetRequest) throws Exception {
+
+        log.info(" > > > POST /api/v1/auth/reset (Requesting password reset link)");
+
+        passwordResetService.getPasswordResetLink(passwordResetRequest);
+
+        log.info(" < < < POST /api/v1/auth/reset (Requesting password reset link successful)");
+
+        return ResponseEntity.ok("Password reset link sent successfully");
+    }
+
+    /**
+     * Save a new password request POST > /api/v1/auth/reset/password
+     */
+    @PostMapping("/reset/password")
+    public ResponseEntity<Object> saveNewPassword(@Validated(PasswordResetValidator.saveNewPassword.class)
+                                                  @RequestBody PasswordResetRequest passwordResetRequest) throws Exception {
+
+        log.info(" > > > POST /api/v1/auth/reset/password (Updating user's password)");
+
+        passwordResetService.saveNewPassword(passwordResetRequest);
+
+        log.info(" < < < POST /api/v1/auth/reset/password (Updating user's password successful)");
+
+        return ResponseEntity.ok("Password has been updated successfully");
     }
 }
